Too much for one commit, but inseparable at this point.  Includes: localizing shared constants to shared/i18n, shortening names, decreasing non-debug output, closing file pointers when done, indenting changes, removing API keys from public code, and being explicit about language code formats.

+++ kalite/i18n/management/commands/generate_subtitle_map.py
@@ -6,6 +6,8 @@ have been subtitled in. This data is then used by the command cache_subtitles
 to intelligently request srt files from Amara's API, rather than blindly requesting
 tons of srt files that don't exist. This command should be run infrequently, but
 regularly, to ensure we are at least putting in requests for the srts that exist.
+
+NOTE: srt map deals with amara, so uses ietf codes (e.g. en-us).
 """
 
 import datetime
@@ -21,17 +23,10 @@ from django.core.management import call_command
 
 import settings
 from settings import LOG as logging
+from shared.i18n import AMARA_HEADERS, SRTS_JSON_FILEPATH
+from shared.i18n import get_language_name, get_lang_map_filepath
 from shared.topic_tools import get_node_cache
 from utils.general import convert_date_input, ensure_dir, make_request
-
-
-headers = {
-    "X-api-username": "kalite",
-    "X-apikey": "9931eb077687297823e8a23fd6c2bfafae25c543",
-}
-
-SRTS_JSON_FILENAME = "srts_remote_availability.json"
-LANGUAGE_SRT_SUFFIX = "_download_status.json"
 
 
 class OutDatedSchema(Exception):
@@ -39,7 +34,7 @@ class OutDatedSchema(Exception):
         return "The current data schema is outdated and doesn't store the important bits. Please run 'generate_subtitles_map.py -f' to generate a totally new file and the correct schema."
 
 
-def create_all_mappings(force=False, frequency_to_save=100, response_to_check=None, date_to_check=None):
+def create_all_mappings(force=False, frequency_to_save=100, response_to_check=None, date_to_check=None, map_file=SRTS_JSON_FILEPATH):
     """
     Write or update JSON file that maps from YouTube ID to Amara code and languages available.
 
@@ -52,15 +47,13 @@ def create_all_mappings(force=False, frequency_to_save=100, response_to_check=No
     videos = get_node_cache('Video').values()
 
     # Initialize the data
-    out_file = settings.SUBTITLES_DATA_ROOT + SRTS_JSON_FILENAME
-
-    if not os.path.exists(out_file):
-        ensure_dir(os.path.dirname(out_file))
+    if not os.path.exists(map_file):
+        ensure_dir(os.path.dirname(map_file))
         srts_dict = {}
     else:
         # Open the file, read, and clean out old videos.
         try:
-            with open(out_file, "r") as fp:
+            with open(map_file, "r") as fp:
                 srts_dict = json.load(fp)
         except Exception as e:
             logging.error("JSON file corrupted, using empty json and starting from scratch (%s)" % e)
@@ -118,13 +111,13 @@ def create_all_mappings(force=False, frequency_to_save=100, response_to_check=No
             continue
 
         if n_new_entries % frequency_to_save == 0:
-            logging.info("On loop %d dumping dictionary into %s" % (n_new_entries, out_file))
-            with open(out_file, 'wb') as fp:
+            logging.info("On loop %d dumping dictionary into %s" % (n_new_entries, map_file))
+            with open(map_file, 'wb') as fp:
                 json.dump(srts_dict, fp)
         n_new_entries += 1
 
     # Finished the loop: save and report
-    with open(out_file, 'wb') as fp:
+    with open(map_file, 'wb') as fp:
         json.dump(srts_dict, fp)
     if n_failures == 0:
         logging.info("Great success! Stored %d fresh entries, %d total." % (n_new_entries, len(srts_dict)))
@@ -142,10 +135,12 @@ def update_video_entry(youtube_id, entry={}):
                             "last_attempt": "2013-07-06",
                         }
     To update an entry, pass it in.
+
+    Note: language_codes are in IETF format (e.g. en-US)
     """
     request_url = "https://www.amara.org/api2/partners/videos/?format=json&video_url=http://www.youtube.com/watch?v=%s" % (
         youtube_id)
-    r = make_request(headers, request_url)
+    r = make_request(AMARA_HEADERS, request_url)
     # add api response first to prevent empty json on errors
     entry["last_attempt"] = unicode(datetime.datetime.now().date())
 
@@ -196,17 +191,21 @@ def update_video_entry(youtube_id, entry={}):
         return entry
 
 
-def update_language_srt_map():
+def update_language_srt_map(map_file=SRTS_JSON_FILEPATH):
     """
     Translate the srts_remote_availability dictionary into language specific files
     that can be used by the cache_subtitles command.
+
+    Note: srt map deals with amara, so uses ietf codes (e.g. en-us)
     """
     # Load the current download status
     try:
-        api_info_map = json.loads(open(settings.SUBTITLES_DATA_ROOT + SRTS_JSON_FILENAME).read())
+        with open(map_file) as fp:
+            api_info_map = json.load(fp)
     except Exception as e:
         # Must be corrupted; start from scratch!
-        logging.warn("Could not open %s for updates; starting from scratch.  Error=%s" % (srt_download_info_filepath, e))
+        logging.warn("Could not open %s for updates; starting from scratch.  Error=%s" % (map_file, e))
+        import pdb; pdb.set_trace()
         api_info_map = {}
 
     # Next we want to iterate through those and create a big srt dictionary organized by language code
@@ -234,7 +233,8 @@ def update_language_srt_map():
             lang_map = {}
         else:
             try:
-                lang_map = json.loads(open(lang_map_filepath).read())
+                with open(lang_map_filepath, "r") as fp:
+                    lang_map = json.load(fp)
             except Exception as e:
                 logging.error("Language download status mapping for (%s) is corrupted (%s), rewriting it." % (lang_code, e))
                 lang_map = {}
@@ -272,19 +272,23 @@ def update_language_srt_map():
         remote_availability_map[lang_code].update(lang_map)
 
     # Finally, remove any files not found in the current map at all.
-    for filename in os.listdir(os.path.dirname(lang_map_filepath)):
-        lang_code = lang_code = filename.split("_")[0]
-        if not lang_code in remote_availability_map:
-            file_to_remove = get_lang_map_filepath(lang_code)
-            logging.info("Subtitle support for %s has been terminated; removing." % lang_code)
-            if os.path.exists(file_to_remove):
-                os.remove(file_to_remove)
-            else:
-                logging.warn("Subtitles metadata for %s not found; skipping deletion of non-existent file %s." % (lang_code, file_to_remove))
+    if lang_map_filepath:
+        for filename in os.listdir(os.path.dirname(lang_map_filepath)):
+            lang_code = lang_code = filename.split("_")[0]
+            if not lang_code in remote_availability_map:
+                file_to_remove = get_lang_map_filepath(lang_code)
+                logging.info("Subtitle support for %s has been terminated; removing." % lang_code)
+                if os.path.exists(file_to_remove):
+                    os.remove(file_to_remove)
+                else:
+                    logging.warn("Subtitles metadata for %s not found; skipping deletion of non-existent file %s." % (lang_code, file_to_remove))
+
     return remote_availability_map
 
 
 def print_language_availability_table(language_srt_map):
+    """srt map deals with amara, so uses ietf codes (e.g. en-us)"""
+
     logging.info("=============================================")
     logging.info("=\tLanguage\t=\tNum Videos\t=")
     for lang_code in sorted(language_srt_map.keys()):
@@ -292,13 +296,14 @@ def print_language_availability_table(language_srt_map):
     logging.info("=============================================")
 
     n_srts = sum([len(dict) for dict in language_srt_map.values()])
-    logging.info("Great success! Subtitles support found for %d languages, %d total dubbings!" % (len(language_srt_map), n_srts))
-
-def get_lang_map_filepath(lang_code):
-    return settings.SUBTITLES_DATA_ROOT + "languages/" + lang_code + LANGUAGE_SRT_SUFFIX
+    logging.info("Great success! Subtitles support found for %d languages, %d total dubbings!" % (
+        len(language_srt_map), n_srts,
+    ))
 
 class Command(BaseCommand):
-    help = "Update the mapping of subtitles available by language for each video. Location: %s" % (settings.SUBTITLES_DATA_ROOT + "<lang_code>" + LANGUAGE_SRT_SUFFIX)
+    help = "Update the mapping of subtitles available by language for each video. Location: %s" % (
+        get_lang_map_filepath("<lang_code>"),
+    )
 
     option_list = BaseCommand.option_list + (
         # Basic options
